# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SiliconSentinel orchestrator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/uart/audit": {
      "post": {
        "summary": "Uart Audit",
        "operationId": "uart_audit_api_uart_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UartAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/target/profile": {
      "get": {
        "summary": "Profile",
        "operationId": "profile_api_v1_target_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/target/status": {
      "get": {
        "summary": "Status",
        "operationId": "status_api_v1_target_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evidence/{evidence_id}/raw": {
      "get": {
        "summary": "Raw Evidence",
        "operationId": "raw_evidence_api_v1_evidence__evidence_id__raw_get",
        "parameters": [
          {
            "name": "evidence_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evidence Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uart/status": {
      "get": {
        "summary": "Uart Status",
        "operationId": "uart_status_api_v1_uart_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uart/audit": {
      "post": {
        "summary": "Uart Audit",
        "operationId": "uart_audit_api_v1_uart_audit_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UartAuditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "summary": "Session",
        "operationId": "session_api_v1_sessions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/read": {
      "post": {
        "summary": "Read",
        "operationId": "read_api_v1_memory_read_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryReadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registers/read": {
      "post": {
        "summary": "Registers",
        "operationId": "registers_api_v1_registers_read_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshots/capture": {
      "post": {
        "summary": "Snapshot",
        "operationId": "snapshot_api_v1_snapshots_capture_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/target/halt": {
      "post": {
        "summary": "Halt",
        "operationId": "halt_api_v1_target_halt_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/target/resume": {
      "post": {
        "summary": "Resume",
        "operationId": "resume_api_v1_target_resume_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/target/step": {
      "post": {
        "summary": "Step",
        "operationId": "step_api_v1_target_step_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/write": {
      "post": {
        "summary": "Write",
        "operationId": "write_api_v1_memory_write_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TargetProfile": {
        "properties": {
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "architecture": {
            "type": "string",
            "const": "arm",
            "title": "Architecture",
            "default": "arm"
          },
          "endianness": {
            "type": "string",
            "enum": [
              "little",
              "big"
            ],
            "title": "Endianness",
            "default": "little"
          },
          "address_width": {
            "type": "integer",
            "const": 32,
            "title": "Address Width",
            "default": 32
          },
          "address_spaces": {
            "items": {
              "type": "string",
              "enum": [
                "virtual",
                "physical"
              ]
            },
            "type": "array",
            "title": "Address Spaces",
            "default": [
              "virtual"
            ]
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/Region"
            },
            "type": "array",
            "title": "Regions"
          },
          "registers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Registers",
            "default": [
              "pc",
              "lr",
              "sp",
              "cpsr"
            ]
          },
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "read",
                "registers",
                "snapshot",
                "halt",
                "resume",
                "step",
                "write"
              ]
            },
            "type": "array",
            "title": "Capabilities",
            "default": [
              "read",
              "registers",
              "snapshot"
            ]
          },
          "provenance": {
            "type": "string",
            "title": "Provenance"
          },
          "verified_for_live": {
            "type": "boolean",
            "title": "Verified For Live",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_id",
          "regions",
          "provenance"
        ],
        "title": "TargetProfile"
      },
...
      "UartAuditRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "observe",
              "interrupt",
              "credential_check"
            ],
            "title": "Mode",
            "default": "observe"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 10.0,
            "title": "Timeout Seconds",
            "default": 90
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UartAuditRequest"
      },
...
      "EdgeStatus": {
        "properties": {
          "schema_name": {
            "type": "string",
            "const": "jtagent.edge.status",
            "title": "Schema Name"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "application": {
            "type": "string",
            "const": "ok",
            "title": "Application"
          },
          "build": {
            "$ref": "#/components/schemas/BuildIdentity"
          },
          "source_drift": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Drift"
          },
          "debugger": {
            "$ref": "#/components/schemas/DebuggerStatus"
          },
          "target": {
            "$ref": "#/components/schemas/TargetStatus"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "target_backend": {
            "type": "string",
            "enum": [
              "mock",
              "replay",
              "openocd"
            ],
            "title": "Target Backend"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation"
          },
          "capabilities": {
            "items": {
              "type": "string",
              "enum": [
                "read",
                "registers",
                "snapshot",
                "halt",
                "resume",
                "step",
                "write"
              ]
            },
            "type": "array",
            "title": "Capabilities"
          },
          "recovery_required": {
            "type": "boolean",
            "title": "Recovery Required"
          },
          "snapshots_armed": {
            "type": "boolean",
            "title": "Snapshots Armed"
          },
          "retains_raw_locally": {
            "type": "boolean",
            "title": "Retains Raw Locally"
          },
          "read_prerequisite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read Prerequisite"
          },
          "live_operations": {
            "$ref": "#/components/schemas/LiveGate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_name",
          "schema_version",
          "observed_at",
          "application",
          "build",
          "debugger",
          "target",
          "target_backend",
          "generation",
          "capabilities",
          "recovery_required",
          "snapshots_armed",
          "retains_raw_locally",
          "live_operations"
        ],
        "title": "EdgeStatus",
        "description": "Edge status v2. The unversioned flat shape (state, target_backend, ...) is schema v1."
      },
...
      "MemoryReadRequest": {
        "properties": {
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "expected_generation": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Generation"
          },
          "address": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Address"
          },
          "length": {
            "type": "integer",
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Length",
            "default": 256
          },
          "address_space": {
            "type": "string",
            "enum": [
              "virtual",
              "physical"
            ],
            "title": "Address Space",
            "default": "virtual"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_id",
          "session_id",
          "address"
        ],
        "title": "MemoryReadRequest"
      },
...
      "SessionRequest": {
        "properties": {
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "operations": {
            "items": {
              "type": "string",
              "enum": [
                "read",
                "registers",
                "snapshot",
                "halt",
                "resume",
                "step",
                "write"
              ]
            },
            "type": "array",
            "title": "Operations",
            "default": [
              "read",
              "registers",
              "snapshot"
            ]
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/Region"
            },
            "type": "array",
            "title": "Regions"
          },
          "byte_budget": {
            "type": "integer",
            "maximum": 16384.0,
            "minimum": 1.0,
            "title": "Byte Budget",
            "default": 16384
          },
          "operation_limit": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 1.0,
            "title": "Operation Limit",
            "default": 24
          },
          "ttl_seconds": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Ttl Seconds",
            "default": 120
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_id",
          "regions"
        ],
        "title": "SessionRequest"
      },
...
      "TargetRequest": {
        "properties": {
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "expected_generation": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Generation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_id",
          "session_id"
        ],
        "title": "TargetRequest"
      },
      "ToolResult": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "duration_ms": {
            "type": "number",
            "title": "Duration Ms"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "provenance": {
            "type": "string",
            "title": "Provenance"
          },
          "outcome_uncertain": {
            "type": "boolean",
            "title": "Outcome Uncertain",
            "default": false
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "success",
          "request_id",
          "duration_ms",
          "provenance"
        ],
        "title": "ToolResult"
      },
...
    }
  }
}